+++ firestorm_flutter/lib/gen/firestorm_builder.dart
@@ -93,6 +93,9 @@ class FirestormBuilder implements Builder {
       ImportGenerator.generateImports(importsBuffer, assetID);
     }
 
+    //Generate import for firestore library:
+    ImportGenerator.generateImportFromString(importsBuffer, "package:cloud_firestore/cloud_firestore.dart");
+
     //Generate the converter functions
     RegistryGenerator.generateConverterFunctions(converterBuffer, allFilesClassHolder);