+++ firestorm_flutter/lib/gen/import_generator.dart
@@ -6,4 +6,10 @@ class ImportGenerator {
     importsBuffer.writeln("import '${asset.uri}';");
     return importsBuffer.toString();
   }
+
+  static String generateImportFromString(StringBuffer importsBuffer, String importPath) {
+    importsBuffer.writeln("import '$importPath';");
+    return importsBuffer.toString();
+  }
+
 }